backend/requirements.txt
Django==4.2.18
django-cors-headers==4.3.1
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.196
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3 